Changed the IP of frigg - the entire block is on the Spamhaus Policy Block List (PBL).
This *should* be okay if people didn't misused this list (being listed is simply meant
to mean that you won't connect directly (I send via senddrid, as do most GCE instances).

That said, some people simply overlbock without understanding and don't want to listen
to any form of reason.  After all, I can't force you to receive me email and if you
don't want to, even if it's a misunderstanding of a technical issue, then so be it.